third_party/blink/web_tests/http/tests/websocket/bad-sub-protocol-empty-expected.txt

Test WebSocket bad sub-protocol names (empty).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception SyntaxError: Failed to construct 'WebSocket': The subprotocol '' is invalid..
PASS successfullyParsed is true

TEST COMPLETE
